Write parser optional method. Required changes to ParserResult and Either.
In ParserResult provide a way for optional to check if successful, and to get data on success.
In Either, provide way for ParserResult to check for right-handedness, and to get right-handed data out of an Either. (i,.e., the Either operations that ParserResult uses under the hood).

A few lines of code in main.ts showing an optional parser.